Time: 0 ms (100%), Space: 15.7 MB (67.61%) - LeetHub

+++ 2341-count-prefixes-of-a-given-string/2341-count-prefixes-of-a-given-string.cpp
@@ -7,9 +7,9 @@ public:
            string str = words[j];
            int i  = 0;
            bool good = true;
-        while(i < (int)str.size()){
+        while(i < str.size()){
             
-            if(i>= (int)s.size() || str[i] != s[i]){
+            if(i>= s.size() || str[i] != s[i]){
                 good = false;
                 break;
             }
